Cloud Functions/ingest-shelterluv-email/requirements.txt
functions-framework==3.*
google-cloud-firestore
requests
pytz
pandas
openpyxl